Converting Routing and Dialogs to Bloc
Initially, the LoadingDialog was added and it will be called during the log-in process. Some adjustments were made in AuthEvent and AuthState, such as: adding SendEmailVerication and Register events; adding Unitialized state, and updating LoggedOut by adding the isLoading property and working with Equatable. In LoginView and RegisterView, adjustments like wrapping the build's return with a BlocListener and handling their exceptions in listener were made too. Finally, most of the routes were removed, since the routing in LoginView and RegisterView is now managed by AuthBloc.