Adding LoadingScreenController and fixes
The LoadingScreenController will control loading screens that may appear in the application (at this point, it only shows up in LoginView and Main). The AuthState class has two new properties called 'isLoading' and 'loadingText', allowing all subclasses to provide a LoadingScreen if necessary. Also the main.dart now returns a BlocConsumer, since it will check if the state is loading to show or hide the LoadingScreen.